pyproject.toml
[tool.poetry]
name = "finalproject-kolesnik-m25-555"
version = "0.1.0"
description = "Консольное приложение валютного кошелька"
authors = ["Колесник Матвей Анатольевич"]
license = "MIT"
readme = "README.md"
packages = [{include = "valutatrade_hub"}]

[tool.poetry.dependencies]
python = "^3.12"
prettytable = "^3.17.0"
python-dotenv = "^1.2.1"
requests = "^2.32.5"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.8"

[tool.poetry.scripts]
project = "valutatrade_hub.cli.interface:main"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
